face_recognition_backend/requirements.txt
fastapi
uvicorn
python-dotenv
numpy
opencv-python
face_recognition
Pillow
python-multipart
SQLAlchemy